TCPClient: Extra debug scenarios added. TODO ...
The DK needs to be notified if the remote closes the socket ... or if the connect() actually resulted in a connection beyond the gateway, for that matter. Currently, a TCP connection success is concluded merely from getting the request out to the gateway. This needs to be only if the gateway was in fact able to make the TCP connection on our behalf!